a function that creates an array of integers.

+++ more_malloc_free/3-array_range.c
@@ -17,7 +17,7 @@ int *array_range(int min, int max)
 	if (min > max)
 		return (NULL);
 	/*allocation*/
-	pArray = (int)malloc(max * sizeof(int));
+	pArray = malloc(max * sizeof(int));
 	if (pArray == NULL)
 	{
 		free(pArray);
